requirements.txt
fastapi
uvicorn
groq
python-dotenv
streamlit
requests